backend/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.0
<<<<<<< HEAD
PyJWT==2.8.0 
=======
PyJWT==2.8.0 
gtts==2.5.4
googletrans==4.0.0rc1
raftos==0.6.0
>>>>>>> RAFT-2
